Fix mocha-to-qunit issues
* Modules with nested modules were being skipped when setting up
callback hooks. I fixed that by changing the setupCallbackHooks function
in `cleanup-imports.js` to look for all of the paths of the callback
functions in the root. By doing that, I could append `hooks.` to the
callee of the closest ExpressionStatement, which would be the
callback, and set the `hooks` parameter in the closest `module`
CallExpression. Now all modules with callback hooks will be set up
properly.
* Some expect statements were being changed to use non-standard assert
methods that are provided in another library. I fixed those transforms
to use standard assert methods with some workarounds.
* Test fixtures had to be updated to match the standard assert methods.